api/requirements.txt
# ChartSense API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.0  # PostgreSQL support for production

# Trading - Alpaca
alpaca-py>=0.13.0

# Data Analysis
numpy>=1.26.0
pandas>=2.1.0

# Scheduling
apscheduler>=3.10.0

# Timezone handling
pytz>=2024.1

# AI/LLM (optional)
openai>=1.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=24.1.0
ruff>=0.1.0
